Move placeholder node mapping out of links
